Add files via upload
The break and continue statements are used to control the flow of loops in programming. The break statement is used to immediately terminate the loop, regardless of the iteration or condition. When break is encountered, the control exits the loop entirely, skipping any remaining iterations. It is typically used when a specific condition is met, and further execution of the loop is no longer required. On the other hand, the continue statement is used to skip the current iteration and proceed directly to the next one continue provide flexibility in managing loops efficiently and are essential tools for handling complex flow control scenarios.