Add files via upload
HashSet in Java is a collection that implements the Set interface and does not allow duplicate elements. It is backed by a hash table, providing fast access to elements, with an average time complexity of O(1) for basic operations like add, remove, and contains. HashSet stores elements in an unordered manner, meaning the order of elements is not guaranteed. It is ideal for use cases where the uniqueness of elements is important and the order of insertion is not. Features of HashSet include null element support, efficient lookups, and the ability to perform set operations like union, intersection, and difference.