Add files via upload
A HashMap in Java is a part of the java.util package and implements the Map interface. It stores data in key-value pairs, where each key is unique. It allows for fast retrieval, insertion, and deletion operations, with an average time complexity of O(1) for these operations. The HashMap uses a hashing mechanism to store data, making it efficient for large datasets. It allows for the mapping of objects to values, including primitive data types, and handles null values as well. It's commonly used for tasks like counting occurrences, caching data, and organizing key-value pairs.