Add files via upload
The PASCAL_BINOIAL class computes and prints the values from the last row of Pascal's Triangle for a given number n.

fact(int n): This method calculates the factorial of a number n by multiplying all integers from 1 to n.

binomial_coefficient(int r, int c): This method calculates the binomial coefficient 
𝐶
(
𝑟
,
𝑐
)
=
𝑟
!
𝑐
!
(
𝑟
−
𝑐
)
!
C(r,c)= 
c!(r−c)!
r!
​
 , which is used in Pascal's Triangle.

printlast(int n): This method prints the values of the last row of Pascal's Triangle for a given n. It uses the binomial coefficient for each index i from 0 to n-1.

For n = 4, the output is 1 3 3 1, which is the fourth row of Pascal's Triangle (starting from row 0).