Add files via upload
he Power_of_Any_Number class implements several algorithms to perform various tasks on arrays:

power_of_any(int num): This method checks whether a number is divisible by 6 repeatedly until it is no longer divisible. If the result is 1, the number is a power of 6; otherwise, it's not.

missing_Number(): This method finds the missing number in an array of consecutive integers. It sorts the array and compares each element with its index. If they don't match, the missing number is identified.

greater_tnext(): This method finds elements in the array where the sum of the current and next elements is greater than the third element ahead. The result is stored in an ArrayList.

oddeven_differences(): This method calculates the difference between the sum of odd and even numbers in the array.

totalcontains(): This method calculates how many times each element in the array can be divided by 12 and sums the results.

single_number(): This method uses XOR to find the number that appears only once in an array of numbers.

twosum(): This method finds two elements in the array whose sum equals a target value. It uses a HashMap to store the difference between the target and each element